Focus ring on listbox button when it's open, no ring on dropdown
This approach (having the ring on the button on `focus` or
`group-data-open`) creates a slight flicker when the dropdown closes,
I think because there's an instant where the `open` data attribute is
gone but the button isn't focused yet. It's not terrible, so I'm
satisfied with this state.

I looked into using `peer` but it doesn't work; the element whose
focus state we need to inspect (the `<Options>`) isn't actually a peer
because it's inside the `<Transition>`.