Add email-required attribute to calculator
## Description

A new boolean attribute that requires the email field in the top form
to be filled in if it's shown. (It already uses the browser's
validation for email address well-formedness.)

This has no effect on the mini email signup form that gets shown in
the empty-results state.

As far as I know currently, we don't need to send anything additional
to the email-signup endpoint; if it turns out we do, I'll do that as a
followup PR.

https://app.asana.com/0/1203194244067918/1206625793385634/f

## Test Plan

- Delete the `RA-calc-email-submitted` key from local storage if it's there.

- With neither `show-email` nor `email-required`, make sure you can submit the form.

- With only `show-email`, make sure you can submit the form without entering an email.

- With only `email-required`, make sure the email field is not shown and you can submit the form.

- With both, make sure the email field is shown, without the "(optional)" in the label, and trying to submit the form without an email results in a "please fill in this field" prompt from the browser, and the form isn't submitted.

- Fill in an email, and make sure the form submits and the email field disappears.

- Make sure the form is still submittable.