Fix broken es.xlf; test for this in CI (#135)
## Description

Running `strings:extract` on main is broken because the XML in es.xlf
is broken, so this fixes that.

To make sure this doesn't happen again, and to protect against some
other failure modes like forgetting to run `strings:build`, I'm adding
a check to CI that runs extract, runs build, and asserts that no
changes have been made.

The switch to the `rexreplace` package instead of sed is because sed's
arguments are different on macOS and Linux (BSD vs. GNU), and that
command now has to be runnable on both.

In the long run, I want to use a localization platform, which will
then become the source of truth; that will mean getting rid of es.xlf
and thus this kinda hacky workflow.

## Test Plan

Tested locally as best I could with github.com/nektos/act , but we'll
see what happens on actual GitHub.

- Without fixing es.xlf file, run the workflow and make sure it fails.

- Add a new `msg()` call without running `strings:extract`, commit,
  make sure the workflow fails.

- Add a new `msg()` call, run `strings:extract`, but don't run
  `strings:build`; commit, make sure the workflow fails.

- With the es.xlf fix and no new strings, make sure the workflow
succeeds.