Fix improper initialization of custom HTML element
## Description

Unbeknownst to me, and apparently to many authors of publicized
shadow-DOM example code, you're not supposed to modify the DOM at all
in the constructor of a custom element; you have to do it in
`connectedCallback`.

This bug manifested as the calculator not rendering, and a console
error `Failed to construct 'CustomElement'`. As far as I've found, it
happens if both of these are true:

- The script tag that imports the bundled `state-calculator.js` file
  must _not_ have `type="module"`

- The script tag has to come _before_ the
  `rewiring-america-state-calculator` tag in the DOM

The first one is not true on the embed demo site, and the second one
is not true on `homes.rewiringamerica.org`, which is why we didn't
catch it.

## Test Plan

I reproed the bug by building for prod without this change, and adding
a little test page to `dist`:

```html
<html>
  <head>
    <title>Embed test</title>
  </head>
  <body>
    <script src="/state-calculator.js"></script>
    <link rel="stylesheet" type="text/css" href="/rewiring-fonts.css" />
    <rewiring-america-state-calculator api-key="redacted">
    </rewiring-america-state-calculator>
  </body>
</html>
```

I served the directory using the `serve` package, and made sure the
calculator does not show up. Moving the script tag below the
calculator element, or adding `type="module"`, makes the calculator
show up.

After applying this change and rebuilding, the calculator shows up on
the test page as above. Try the other configurations (adding
`type="module"`, moving the tag around) to make sure those work too.

Modify the calculator's `lang` attribute to make sure that still works
(since the initialization logic of it has changed a bit.)

To try to replicate the conditions more easily, I tried removing
`type="module"` from the tag on state-calculator.html, but Parcel
won't let you do that.