Send email-required flag to signup endpoint
## Description

We'll want to track this along with signups to measure the effect of
requiring email. This will require a followup to Zuplo to read this
param and send it to SFMC.

## Test Plan

Sign up under both conditions, make sure the flag is present and with
the correct value in network inspector. (Zuplo isn't reading the flag
yet so nothing will change in SFMC.)