Fix the tax credit/rebate messaging on old embed (#148)
## Description

The old embed is not long for the world, but it's currently saying
some incorrect things, so best fix them.

The text and conditions are copied from the old IRA calculator on the
main site. That frontend actually displays one other notice, only to
renters who are under 150% AMI. I'm not reproducing that here, again
because this frontend is on the way out, and it's not crucial info.

https://app.asana.com/0/1206661332626418/1206707343301439

## Test Plan

On the old embed, look up incentives for 80212 and $2000; make sure
the "you may not qualify for tax credits" message appears. Look up
incentives for 80252 and $200k; make sure the "your income is above
the threshold" message appears. Do $80k and make sure you get no
message (because you get tax savings and you're under 150% AMI).