Use start dates from API for "expected in" message
## Description

Having finally gotten our dates story straight on the API side, we can
now just use them here in the frontend.

The old frontend just gets years from v0, so it displays those as-is
if they're in the future. The new frontend determines whether the
pseudo-ISO-8601 date is in the future, and displays the year from it
if so. (In the interest of time, I didn't want to go further in this
PR by rendering "2024H2" as "late 2024" or whatever, though we can
easily go that route later.)

I also added Jest setup so I could write a unit test for the date
logic. I don't know how much other unit testing we'll need in this
codebase, but now the infrastructure is there for when we want it.

## Test Plan

New unit test for the date parsing. (Useful, because JS's Date class
is a nightmare to work with!)

Set my local API instance as the API host (`api-host` attribute on the
calculator component). Look at the old frontend and make sure all the
IRA rebates show up with "2025". Look at the new frontend with a
low-ish income, and make sure the IRA rebates show up with "Expected
in 2025", and they are shown after all other incentives in their item
category.