Use more_info_url from v1 API as link button destination
## Description

The links to the RA pages about the IRA incentives are currently in
`item.url`, which doesn't make conceptual sense. We're going to move
them to a new `more_info_url` field on incentives, and make
`program_url` required. We'll have

The backend is currently not sending `more_info_url` in v1 at all, so
this PR will be a no-op when it's deployed. After the backend change
is deployed, I'll follow up by removing `item.url` from the frontend,
and then from the backend.

## Test Plan

Run against the prod backend and make sure the IRA incentives have
link buttons with the same contents and URL as before.

Run against a local backend with the `more_info_url` change applied;
make sure the frontend continues to have exactly the same
behavior. (It used to be that the IRA incentives had no program URL,
so the buttons used the item URL. Now they do have a program URL, but
they all have a more-info URL that is exactly the same as the item
URL.)