Stop using item.url; make program_url required (#152)
## Description

Followup to #150. This will let us remove item URLs from v1 if we want.

## Test Plan

This should be a no-op; the removed case should never be hit now that
program_url is required. Submit the form for a RI zip code and look at
HVAC incentives; make sure the non-IRA ones still say "visit site" and
the IRA ones say "learn more" and link to the homes.rewiring pages.