Add disclaimer about residence location to city/county/other incentives (#153)
## Description

See the code comment for rationale. This is the simplest, bluntest
possible approach. Other possibilities are:

- Include the locality's name in this disclaimer message. This is not
  trivial to do because, although the API returns authority names,
  they're not necessarily suitable for interpolating into a
  sentence. E.g. Denver is named `City and County of Denver`, which
  does not work in an interpolation like `Limited to residents of
  ${name}`. (These names also aren't currently localized.)

  This approach also doesn't work for `other`-type incentives, because
  those authorities' names aren't place names.

- Have the API somehow indicate when a locality match is ambiguous,
  and only show the disclaimer in those cases. This would considerable
  work on the backend, to improve its simplistic mapping of zip to
  city and county name.

But given that we're imminently launching a state with local
incentives, I want to have something in place to cover this case.

The Spanish translation is from Google Translate, but I Googled the
phrase to confirm that very similar phrases are in wide use.

One final note: `other`-type incentives may be utility-based, not
locality-based, but the API doesn't convey that info.

## Test Plan

Search for HVAC incentives in 80212 to get some Denver incentives;
make sure the text shows up on the city incentives but not on others.