Handle multi-item incentives, and more items (#158)
## Description

This adds human-friendly labels for all the new items, plus labels for
a few specific groups of items. (E.g. an incentive for both
`ducted_heat_pump` and `ductless_heat_pump` will be shown as being for
"an air-source heat pump".)

I factored out the name logic into a new file because it's getting
pretty extensive.

The Spanish translations are taken from existing translations of
incentive descriptions (which, helpfully, serves to validate that all
the new items exist in incentives we already have). So I can't fully
vouch for them myself, but they are sourced from our human-vetted
translations.

## Test Plan

Cypress tests pass. New unit test for names of multi-item incentives.

Manual testing of a few queries. The results should still match prod
exactly because none of the new items are used in the backend yet.
(The exception is the non-HP dryers and water heaters -- those are
already being used in the backend so will newly show up in the
frontend with this PR. Looked at those in English and Spanish to make
sure the headline looks good. ZIPs 81130 + Gunnison County Electric
and 81004 + San Isabel Electric)