[WAIT TO MERGE] Don't show generic HEAR info for NY
## Description

This adds a way to suppress the generic HEAR/HER cards for users in
some states. NY about to launch their HEAR program, so this excludes
them from the HEAR card.

Wait to merge this until either we add the detailed EmPower+ program
info to the API, or we specifically decide to remove the generic info
anyway.

## Test Plan

Submit the form with zip 12084 and a low income; make sure there are
no HEAR cards (but there is still a HER card under
Weatherization). Change the zip to 02861 (RI) and make sure there are
HEAR cards.