Include audits in weatherization project; be looser in which incentives to show
## Description

We got a report that a Vermont weatherization incentive wasn't showing
up even though it's in our data; the problem was that its `item`
include both weatherization stuff and `energy_audit.`

Either of these changes would fix the immediate problem of the VT
weatherization incentive not showing up, but they both make sense in
isolation:

- Add `energy_audit` to the list of items included in the
  "weatherization & efficiency" project.

- When you select a project tab, it should show incentives that have
  _any item in common_ with the project, not just ones whose items are
  a _subset_ of the project's items.

## Test Plan

Income of $50k and zip 05401 should make a VT "100% of cost of an
energy audit and weatherization" incentive appear.

The $150 federal tax credit for an audit should also newly appear
(this is fixed by the first bullet point above).