New design for incentive cards
## Description

I believe this is exactly as designed. There are a couple of things
that weren't 100% clear to me:

- The max-width of these cards is 720px; does that mean the max width
  of the _entire calculator_ should also be 720px, including the form?
  (That's what I've implemented here.)

- If the headline is long, it can run into the top-right CTA link
  (especially if the CTA text is "Learn how to apply"). The designs
  didn't specify what should happen in that case. I made the headline
  line-wrap, while the CTA link stays flush against the top and right.
  The CTA link never line-wraps.  There's at least a 4-rem (16px) gap
  between the headline and CTA. Screenshot attached.

## Follow-up

CI will fail because of translations. If copy looks good, do those.

## Test Plan

Look at the cards on desktop (whole variety of window sizes) and
mobile (portrait and landscape). Make sure to look at an IRA rebate
card, which gets two chips that line-wrap on narrow layout.