"Coming soon" card
## Description

In states where we don't have SLU coverage, show this card at the
bottom of the results to let people know there may be more incentives
available that we haven't returned.

### Next steps

Wait for Spanish translation of copy. (Request is already submitted.)

## Test Plan

Search for zip 57104 (SD), select a project, and see the card at the
bottom of the results. This is encoded in a new Cypress test.