add a slot to customize email helper text (#179)
## Description

Allow overriding the helper text on the email field using a slot, if
needed. This allows partners using advanced functionality
(`show-email="true"` and event handlers to facilitate signing up users
to their own email list) to customize the email label so it explains
what's going on. This could be used in conjunction with customizing the
footer as well to further clarify which terms and privacy policy apply.

This is a prototype - seeking feedback to see if this makes sense to
others, pending legal input on the approach as well.

## Test Plan

I'm running `yarn cypress:open` and inspecting that the calculator looks
reasonable when the last of the new email tests runs. The preview build
for this PR should be unchanged from production.

This is a screenshot from the Cypress test:
![Screenshot 2024-07-19 at 5 35
03 PM](https://github.com/user-attachments/assets/741178b9-1ba4-4d36-bd85-0f3f9939b6d7)