Refactor logic for HEAR multi-item incentive names (#188)
## Description

This is what I was talking about in #187.

It ended up being more involved than I thought, because Jest couldn't
handle the JSX in the file `projects.tsx`. I tried other values of the
`jsx` setting in tsconfig.json, but then Jest couldn't handle the
magical `jsx:...` imports in projects.tsx (understandable; that's done
using Parcel magic).

So I factored out the icons from projects.tsx into a new file, and now
the unit test works fine.

## Test Plan

`yarn test`. The unit test is unmodified but still passes.

Look at RI's HEAR incentive for stoves/dryers, and make sure it shows
up and has the correct headline.

Make sure the icons in the projects dropdown are correct. The one in
the dropdown button should be purple, and in the menu should be
grey-700 (i.e. almost black).