Exclude generic HEAR/HER cards for DC (#192)
## Description

DC has rolled out their HEAR/HER rebates. This PR removes the "coming
soon" cards.
[Asana](https://app.asana.com/0/1207456219977120/1208649430166793/f)

This repo also doesn't appear to work with Node v20, and although we
specify 18 in the README, we don't enforce it. This PR adds an `.nvmrc`
to make that easier.

## Test Plan

check Weatherization category for a low income 20002 zip code household
to make sure that the "Expected in 2024-2025" cards don't show up
anymore