Add flag to send email signups to staging instead of prod
## Description

This is a no-op for the moment; Zuplo-side change to read the flag and
act on it is upcoming.

## Test Plan

Watch network traffic in web inspector.

With the flag set, submit an email and make sure the flag is present
and set to true in the payload.

Without the flag set, submit an email and make sure the flag is not
in the payload at all.