Add gas utility selector
## Links

- https://app.asana.com/0/1208668890181682/1208886892124341

## Description

- Show the gas utility selector if the `/utilities` call returns a
  response with the `gas_utilities` key present. This is the way I
  came up with for the API to indicate whether the choice of gas
  utility matters: key is present means yes, key is absent means no.
  If the key is present but there are no gas utilities, we still show
  the selector to offer the choice of "no gas service" vs. "other",
  which is significant in the Mass Save case.

- The selector always has three items at the bottom, after any real
  utilities: "Delivered propane or fuel oil", "No gas service", and
  "Other". The former two behave exactly the same behind the scenes,
  but are intended to reduce confusion: users may not know that we
  mean natural/methane gas pipeline service when we say "gas service"
  or "gas utility".

- I had to add the `previousResponse` field to FetchState's loading
  state, so that the gas utility selector can stay visible and
  populated if you enter a new zip code. Otherwise, the selector would
  disappear while the `/utilities` call was pending, and it would look
  janky.

As implemented, when you enter a zip code in MA and the `/utilities`
call completes, the gas utility selector just pops in. This is pretty
disruptive to the rest of the form in the two-column view, because all
the other fields reposition --- they're in the same tabbing order, but
ones that were on the left are now on the right and vice versa. It's
not so bad in one-column view. I think this is tolerable for now,
especially since the selector will only appear for MA zip codes. But
it may be worth a rethink.

## Test Plan

Add `api-host` attribute to the element on index.html to hit my local
API server (so gas utilities are returned).

Enter zip codes in MA to see the gas selector pop in, and make sure
the options make sense. E.g. 02130 (Boston; has Eversource) and 01011
(middle of nowhere; has no gas).

Enter one MA zip code, then another; make sure the gas selector stays
visible throughout. Then enter a non-MA zip code; make sure the
selector disappears once the new location's electric utilities are
populated.

Submit the form with the various gas options selected. Use network
inspector to check the `gas_utility` param sent to the API. (Currently
the API's response won't change based on it.) It should be `none` if
"delivered fuels" or "no gas service" is selected, absent if "other"
is selected, and a utility ID if a real gas utility is selected.