Include beta states by default (#200)
## Description

People using the embed seem to forget to enable the beta state option
fairly often. Let's set it to be on by default since the main use of
this page is testing the calculator with pre-release data.

## Test Plan

Load the preview build and verify that the beta checkbox is selected and
that `include-beta-states` is set on the embed element. Ensure that
`include_beta_states` is passed to the API when you submit the
calculator. Disable the checkbox and verify that `include-beta-states`
is removed from the widget and that `include_beta_states` is not set
when you submit the calculator.