Actually send gas utility to API
## Description

If you chose an actual gas utility, as opposed to one of the "no
service" or "other" items, we just weren't sending it. Oooops

## Test Plan

Enter a MA zip (01945), choose "National Grid" in the gas utility
dropdown, and submit; make sure the param is set in the network
inspector. Even if you choose a non-Mass Save electric utility, Mass
Save incentives should come back.