Add link to Switch is On for CA and WA
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209198275861169)

## Description

If the user's entered location is in those states but the API has no
coverage (as indicated by the `coverage` response field), show a
coming-soon message that links to another org that has comprehensive
coverage of those states.

Also make the coming soon card grey. I think it's been that way in
designs for a while but we never got around to changing it?

## Test Plan

Uncheck "include beta states" and enter a CA zip (94306); make sure
the message shows up and the link goes to the Switch is On's CA
page. Same with a WA zip (98281). Enter another non-covered state zip
(82001; WY) and make sure the old message shows up.

Switch to `lang="es"` and repeat the above; make sure the Spanish
message shows up.