Exclude Mass Save from location eligibility message
## Links

- [Asana](https://app.asana.com/0/1208668890181682/1209741993815681)

## Description

This is a short-term fix, in anticipation of later having the API
return more detail regarding eligibility requirements that it can't be
sure you meet. This codebase is, obviously, very much the wrong place
to be conditioning logic on specific authority IDs, but we don't
expect this code to last long. (Famous last words.)

We can add any other utility-only authorities to this set as needed in
the meantime.

## Test Plan

Enter zip 01201 and choose a Mass Save utility. Look under the HVAC
project and make sure the message doesn't appear on the Mass Save
rebates.

Enter zip 80212 and look under the battery project; there are city of
Denver incentives there that should still have the hedge.