Headlineable groups for charger/wiring/panel (#229)
## Description

There are some incentives that combine these.

This requires a tweak to the multi-item naming logic: first try
finding a group for all the items in the incentive, and only if that
doesn't work, filter the items down to those in the currently viewed
project.

The Spanish translations are pretty long but kind of unavoidable.

## Test Plan

Input 02149, National Grid; look at panel & wiring to see one of these
combined incentives. The same one should show up if you look at the
transportation project too.