Fix anchor position after form submit (#240)
## Description

Once I moved the target element up to the state calculator container as
a whole, the "nearest" option for `scroll-if-needed` didn't make sense
anymore, because it could scroll to the bottom of the calculator itself!
I switched to "start" so that the scroll is at the top.

## Test Plan

Scrolled to the bottom of the page (where the form submit is only just
visible at the top) and made sure that it scrolled up to the top of the
summary card after submit, plus additional padding to account for the
floating header at homes.rewiringamerica.org/calculator


https://github.com/user-attachments/assets/02916665-21cc-4cfc-a6d0-f15c8400c6c8