Bill impact calculator, a.k.a. "REM widget" (#252)
## Description

This is a pretty big PR, unfortunately!

I recommend starting the review by running this locally and going to
`localhost:1234/rem.html` to see the flow, and how it progresses through
a series of steps that build up in the UI. (I.e. first there's only a
form, and then it turns into a form snapshot + upgrade selection, and
then it turns into form snapshot + upgrade snapshot + results.)

Most of the changes are new greenfield UI code in `src/rem`. There are a
few changes to common UI components; I've made some inline comments to
explain those.

Tour of the new files in `src/rem`:

- `ApproximateResults.tsx`: the "likely savers" UI; very simple.
- `DetailedResults.tsx`: the "detailed estimate" UI. There's a lot of
logic to determine precisely which version of the copy to show in the
estimates.
- `NumberLine.tsx`: the component that graphically shows the p20,
median, and p80 of the cost and emissions estimates. Some very
ticky-tack UI layout in here.
- `element.tsx`: where the top-level custom element is defined, where
the overall flow of the widget is managed, and where the API call is
made from.
- `form-snapshot.tsx`: the form snapshot. It's visually similar to the
incentives embed's equivalent, but different enough that it didn't feel
worthwhile to make the code shareable. I'm open to revisiting that,
though.
- `form.tsx`: the form. Much simpler than the incentives form since
there's no API call in the middle of it.
- `upgrade-options.tsx`: displays the cards for each available upgrade.
Also very simple.

Notable missing pieces:

- Integrating the REM API's new `estimate_type` flag to choose between
the detailed and approximate result paths.
- Adding a machine-friendly way to communicate error states to the API,
and integrating that here. (This is what will allow us to get rid of the
provisional error state that's implemented here.)
- Cypress or snapshot testing.
- Docs for embedders.

## Test Plan

- Try various addresses and upgrades.
- Try a commercial address, or a nonsense address, to get error cases.
- Try the incentives embed briefly to make sure that's not broken.